+++ src/main/java/com/github/beemerwt/mcrpg/config/ability/SuperBreakerConfig.java
@@ -8,6 +8,6 @@ import com.github.beemerwt.mcrpg.config.SuperAbilityConfig;
 @JanksonObject
 public class SuperBreakerConfig extends SuperAbilityConfig {
     @JankComment("Mining speed multiplier, scaling with skill level")
-    public double baseSpeedMultipier = 1.5; // Mining speed multiplier at level 1
-    public double maxSpeedMultipier = 4.0; // Mining speed multiplier at max level
+    public float baseSpeedMultipier = 1.5f; // Mining speed multiplier at level 1
+    public float maxSpeedMultipier = 4.0f; // Mining speed multiplier at max level
 }
